move eida50 tests into test_eida50
